drywall-takeoff-3d-fbm/drywall-takeoff-3d-fbm/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

google-cloud-storage
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-secret-manager
db-dtypes
vertexai==1.43.0
google-auth
google-oauth
pathlib
pypdf
pdf2image
pydantic
python-multipart
pandas
google-cloud-pubsub
openpyxl
scikit-image
opencv-python
geoip2
roman
numpy
ruamel-yaml
matplotlib
pygltflib
pillow
fastapi
uvicorn